WIP: move instructions into components
